use protobufs and standardized transmitters
