+++ protos/protos.proto
@@ -41,6 +41,8 @@ message AgentLocationUpdatePayload {
     int64 epoch = 1;
     string id = 2;
     string location_id = 3;
+    double location_lat = 4;
+    double location_lon = 5;
 }
 
 message SpaceOccupancyUpdatePayload {
